Rinomina file e rimozione funzioni duplicate
